+++ testing/test.cpp
@@ -1,12 +1,12 @@
 #include "cpp_impl/src/edubot/include/edubot/robot.hpp"
 #include <unistd.h>
 
-constexpr float DEG2RAD = M_PI/180.0;
+constexpr double DEG2RAD = M_PI/180.0;
 
 int main(int argc, char const *argv[])
 {
-    std::vector<float> l = {1, 1, 1, 1};
-    std::vector<float> pos = {50, 115, 180, 50}; 
+    std::vector<double> l = {1, 1, 1, 1};
+    std::vector<double> pos = {50, 115, 180, 50}; 
     Robot robot(l, "/dev/ttyUSB0", 9600, 200);
 
     sleep(2);
